+++ examples/Vulkan/HybridRefractions/VulkanHybridRefractions.cpp
@@ -1181,12 +1181,12 @@ void VulkanHybridRefractions::buildCommonDescriptorSet()
 	lightDataWDS.setBufferInfo(0, pvrvk::DescriptorBufferInfo(_deviceResources->lightDataBuffer, 0, _deviceResources->lightDataBufferView.getDynamicSliceSize()));
 
 	materialBufferWDS.setBufferInfo(0, pvrvk::DescriptorBufferInfo(_deviceResources->materialBuffer, 0, VK_WHOLE_SIZE));
-	for (int i = 0; i < _deviceResources->materialIndexBuffers.size(); i++)
+	for (size_t i = 0; i < _deviceResources->materialIndexBuffers.size(); i++)
 	{
 		materialIndicesWDS.setBufferInfo(i, pvrvk::DescriptorBufferInfo(_deviceResources->materialIndexBuffers[i], 0, _deviceResources->materialIndexBuffers[i]->getSize()));
 	}
 
-	for (int i = 0; i < _deviceResources->textures.size(); i++)
+	for (size_t i = 0; i < _deviceResources->textures.size(); i++)
 	{
 		sampledTexturesWDS.setImageInfo(
 			i, pvrvk::DescriptorImageInfo(_deviceResources->textures[i].imageView, _deviceResources->samplerLinear, pvrvk::ImageLayout::e_SHADER_READ_ONLY_OPTIMAL));
@@ -1194,12 +1194,12 @@ void VulkanHybridRefractions::buildCommonDescriptorSet()
 
 	accelerationStructureWDS.setAccelerationStructureInfo(0, _deviceResources->accelerationStructure.getTopLevelAccelerationStructure());
 
-	for (int i = 0; i < _deviceResources->vertexBuffers.size(); i++)
+	for (size_t i = 0; i < _deviceResources->vertexBuffers.size(); i++)
 	{
 		vertexBufferArrayWDS.setBufferInfo(i, pvrvk::DescriptorBufferInfo(_deviceResources->vertexBuffers[i], 0, _deviceResources->vertexBuffers[i]->getSize()));
 	}
 
-	for (int i = 0; i < _deviceResources->indexBuffers.size(); i++)
+	for (size_t i = 0; i < _deviceResources->indexBuffers.size(); i++)
 	{
 		indexBufferArrayWDS.setBufferInfo(i, pvrvk::DescriptorBufferInfo(_deviceResources->indexBuffers[i], 0, _deviceResources->indexBuffers[i]->getSize()));
 	}
@@ -1826,7 +1826,7 @@ void VulkanHybridRefractions::buildSceneDescriptionBuffer()
 {
 	_deviceResources->_sceneDescription.resize(_vectorModelTransform.size());
 
-	for (int i = 0; i < _vectorModelTransform.size(); ++i)
+	for (size_t i = 0; i < _vectorModelTransform.size(); ++i)
 	{
 		_deviceResources->_sceneDescription[i].modelIndex = 0;
 		_deviceResources->_sceneDescription[i].transform = _vectorModelTransform[i];
@@ -1886,7 +1886,7 @@ void VulkanHybridRefractions::buildModelBuffers(pvrvk::CommandBuffer& uploadCmd)
 		uint32_t numMeshes = _models[j]->getNumMeshes();
 		uint32_t totalIndices = 0;
 
-		for (int meshIdx = 0; meshIdx < numMeshes; meshIdx++)
+		for (uint32_t meshIdx = 0; meshIdx < numMeshes; meshIdx++)
 		{
 			pvr::assets::Mesh mesh = _models[j]->getMesh(meshIdx);
 
@@ -2006,7 +2006,7 @@ void VulkanHybridRefractions::buildSceneElementTransformBuffer()
 	_deviceResources->perMeshTransformBufferView.pointToMappedMemory(_deviceResources->perMeshTransformBuffer->getDeviceMemory()->getMappedData());
 	_deviceResources->perMeshTransformBuffer->setObjectName("PerMeshTransformBuffer");
 
-	for (int i = 0; i < _vectorModelTransform.size(); ++i)
+	for (size_t i = 0; i < _vectorModelTransform.size(); ++i)
 	{
 		_deviceResources->perMeshTransformBufferView.getElementByName("ModelMatrix", 0, i).setValue(_vectorModelTransform[i]);
 	}
@@ -2139,7 +2139,7 @@ void VulkanHybridRefractions::updateScene()
 	_deviceResources->accelerationStructure.buildTopLevelASAndInstances(_deviceResources->device, commandBuffer, _deviceResources->queue,
 		pvrvk::BuildAccelerationStructureFlagsKHR::e_PREFER_FAST_TRACE_BIT_KHR | pvrvk::BuildAccelerationStructureFlagsKHR::e_ALLOW_UPDATE_BIT_KHR, true);
 
-	for (int i = 0; i < _vectorModelTransform.size(); ++i)
+	for (size_t i = 0; i < _vectorModelTransform.size(); ++i)
 	{
 		// Update scene element transforms through the structured buffer view
 		_deviceResources->perMeshTransformBufferView.getElementByName("ModelMatrix", 0, i).setValue(_vectorModelTransform[i]);
